requirements.txt

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0

# Trading and financial analysis
ccxt>=4.0.0
ta>=0.10.2
yfinance>=0.2.0

# Machine learning and AI
scikit-learn>=0.24.2
xgboost>=1.5.0
lightgbm>=3.2.1
joblib>=1.1.0

# Pattern recognition
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# WebSocket and real-time data
websocket-client>=1.3.0
python-socketio>=5.7.0

# Notifications
discord.py>=2.0.0
python-telegram-bot>=20.0.0

# Utilities
python-dotenv>=0.19.0
schedule>=1.1.0
